refactor(web): split Artifact.tsx (3445 lines) into pages/artifact/ modules
One monster file becomes a cohesive directory:
- index.tsx        the Artifact page + data/anchor orchestration
- comment-thread   CommentRow/Card/ResolvedSection/MentionField/Composer/PinnedZone
- comment-panels   OpenPanel + MobileComments
- header-actions   Star/Report/Collections/Tags menus
- insights-history Insights + version history
- rail-deck        Rail/IconBtn/SectionLabel/DeckBar/Presence
- diff-view        DiffView
- types / comment-actions (context) / lib/{layout,markdown,reactions}

No behavior change: typecheck clean, deep artifact spec 6/6 green. Largest file
now 843 lines (was 3445).
